requirements.txt
absl-py==1.3.0
addict==2.4.0
apex
av==10.0.0
cachetools==5.2.0
certifi
charset-normalizer==2.1.1
cPython==0.0.6
cycler==0.11.0
decord==0.6.0
dnspython==2.2.1
dpcpp-cpp-rt==2023.0.0
einops==0.4.1
fonttools==4.38.0
ftfy==6.1.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl==2023.0.0
mkl-fft==1.3.1
mmcv-full==1.6.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pathlib==1.0.1
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sklearn==0.0
tbb==2021.8.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
timm==0.6.7
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchstat==0.0.7
torchvision==0.14.0+cu116
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
zipp==3.9.0
